EnSys-main/requirements.txt
flask
vaderSentiment
spacy
nltk
python-dotenv
scikit-learn
pandas
openai==0.28
python-docx